# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Controle de Estoque API",
        "description": "API para controle de estoque, movimentações e autenticação",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "name": "Leonardo Losso",
            "email": "leonardolosso.sz@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/estoque": {
            "get": {
                "tags": [
                    "Movimentações"
                ],
                "summary": "Listar todos os movimentos de estoque",
                "operationId": "indexStockMovements",
                "responses": {
                    "200": {
                        "description": "Lista de movimentos de estoque",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/StockMovement"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Proibido"
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Movimentações"
                ],
                "summary": "Criar um novo movimento de estoque",
                "operationId": "storeStockMovement",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StockMovement"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Movimento de estoque criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StockMovement"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro na validação dos dados"
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Proibido"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/estoque/{id}": {
            "get": {
                "tags": [
                    "Movimentações"
                ],
                "summary": "Mostrar detalhes de um movimento de estoque",
                "operationId": "showStockMovement",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do movimento de estoque",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do movimento de estoque",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StockMovement"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Movimento não encontrado"
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Proibido"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Movimentações"
                ],
                "summary": "Atualizar um movimento de estoque",
                "operationId": "updateStockMovement",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do movimento de estoque",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StockMovement"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Movimento de estoque atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StockMovement"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro na validação dos dados"
                    },
                    "404": {
                        "description": "Movimento não encontrado"
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Proibido"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/estoque/historico/{productId}": {
            "get": {
                "tags": [
                    "Movimentações"
                ],
                "summary": "Histórico de movimentações de estoque por produto",
                "operationId": "historyByProduct",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "ID do produto para buscar o histórico de movimentações",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Histórico de movimentos de estoque do produto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/StockMovement"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produto não encontrado ou sem histórico"
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Proibido"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "StockMovement": {
                "required": [
                    "product_id",
                    "movement_type",
                    "quantity",
                    "unit_price",
                    "date",
                    "user_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID do movimento de estoque",
                        "type": "integer"
                    },
                    "product_id": {
                        "description": "ID do produto relacionado ao movimento",
                        "type": "integer",
                        "example": 1
                    },
                    "movement_type": {
                        "description": "Tipo de movimento (entrada ou saída)",
                        "type": "string",
                        "example": "entrada"
                    },
                    "quantity": {
                        "description": "Quantidade do movimento",
                        "type": "number",
                        "format": "float",
                        "example": 10
                    },
                    "unit_price": {
                        "description": "Preço unitário do produto no movimento",
                        "type": "number",
                        "format": "float",
                        "example": 15.5
                    },
                    "description": {
                        "description": "Descrição ou observação sobre o movimento",
                        "type": "string",
                        "example": "Compra de mercadoria"
                    },
                    "date": {
                        "description": "Data do movimento",
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-01"
                    },
                    "user_id": {
                        "description": "ID do usuário responsável pelo movimento",
                        "type": "integer",
                        "example": 5
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}